projects/03_python_template/requirements.txt
annotated-types==0.7.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
typing-extensions==4.13.2
urllib3==2.2.3
